requirements.txt
Django==4.1.7
ipdb==0.13.13
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.2
Pillow==9.4.0
django-bootstrap-form
psycopg2-binary
protobuf